removed .long tensor cast

+++ app/SHEPHERD/shepherd/node_embedder_model.py
@@ -170,13 +170,13 @@ class NodeEmbeder(pl.LightningModule):
             # rand_index = torch.tensor(np.vectorize(sample_node_for_et)(all_edge_types.cpu(), targets_dict)).to(device)
             if all_edge_types.numel() == 0:
                 print("No edge types found")
-                rand_index = torch.tensor([], device=device, dtype=torch.long)
+                rand_index = torch.tensor([], device=device, )
             else:
                 rand_indices = []
                 for et in all_edge_types.cpu().numpy():
                     rand_indices.append(sample_node_for_et(et, targets_dict))
 
-                rand_index = torch.tensor(rand_indices, device=device, dtype=torch.long)
+                rand_index = torch.tensor(rand_indices, device=device,)
 
 
             print("Moved: negative sampling")
